paper_trail v0.1.3 PaperTrail.Version

Functions

changeset(model, params \\ :empty)

Creates a changeset based on the model and params.

If no params are provided, an invalid changeset is returned with no validation performed.