PaperTrail.VersionQueries (paper_trail v0.10.1)

Functions

get_current_model(version)

Gets the current model record/struct of a version

get_version(record)

Specs

get_version(record :: Ecto.Schema.t()) :: Ecto.Query.t()

Gets the last version of a record.

A list of options is optional, so you can set for example the :prefix of the query, wich allows you to change between different tenants.

Usage examples:

iex(1)> PaperTrail.VersionQueries.get_version(record, id) iex(1)> PaperTrail.VersionQueries.get_version(record, [prefix: "tenant_id"]) iex(1)> PaperTrail.VersionQueries.get_version(ModelName, id) iex(1)> PaperTrail.VersionQueries.get_version(ModelName, id, [prefix: "tenant_id"])

get_version(model, id)

Specs

get_version(model :: module(), id :: pos_integer()) :: Ecto.Query.t()

get_version(record :: Ecto.Schema.t(), options :: []) :: Ecto.Query.t()

get_version(model, id, options)

Specs

get_version(model :: module(), id :: pos_integer(), options :: []) ::
  Ecto.Query.t()

get_versions(record)

Specs

get_versions(record :: Ecto.Schema.t()) :: Ecto.Query.t()

Gets all the versions of a record.

A list of options is optional, so you can set for example the :prefix of the query, wich allows you to change between different tenants.

Usage examples:

iex(1)> PaperTrail.VersionQueries.get_versions(record) iex(1)> PaperTrail.VersionQueries.get_versions(record, [prefix: "tenant_id"]) iex(1)> PaperTrail.VersionQueries.get_versions(ModelName, id) iex(1)> PaperTrail.VersionQueries.get_versions(ModelName, id, [prefix: "tenant_id"])

get_versions(model, id)

Specs

get_versions(model :: module(), id :: pos_integer()) :: Ecto.Query.t()

get_versions(record :: Ecto.Schema.t(), options :: []) :: Ecto.Query.t()

Gets all the versions of a record given a module and its id

get_versions(model, id, options)

Specs

get_versions(model :: module(), id :: pos_integer(), options :: []) ::
  Ecto.Query.t()