paper_trail v0.8.1 PaperTrail

Functions

delete!(struct, options \\ [origin: nil, meta: nil, originator: nil, prefix: nil])

Same as delete/2 but returns only the model struct or raises if the changeset is invalid.

delete(struct, options \\ [origin: nil, meta: nil, originator: nil, prefix: nil])

Deletes a record from the database with a related version insertion in one transaction

get_current_model(version)

get_model_id(model)

get_version(record)

get_version(model_or_record, id_or_options)

get_version(model, id, options)

get_versions(record)

get_versions(model_or_record, id_or_options)

get_versions(model, id, options)

insert!(changeset, options \\ [origin: nil, meta: nil, originator: nil, prefix: nil])

Same as insert/2 but returns only the model struct or raises if the changeset is invalid.

insert(changeset, options \\ [origin: nil, meta: nil, originator: nil, prefix: nil])

Inserts a record to the database with a related version insertion in one transaction

update!(changeset, options \\ [origin: nil, meta: nil, originator: nil, prefix: nil])

Same as update/2 but returns only the model struct or raises if the changeset is invalid.

update(changeset, options \\ [origin: nil, meta: nil, originator: nil, prefix: nil])

Updates a record from the database with a related version insertion in one transaction