paper_trail v0.8.7 PaperTrail

Functions

delete(struct, options \\ [origin: nil, meta: nil, originator: nil, prefix: nil])

Deletes a record from the database with a related version insertion in one transaction

delete!(struct, options \\ [origin: nil, meta: nil, originator: nil, prefix: nil])

Same as delete/2 but returns only the model struct or raises if the changeset is invalid.

get_current_model(version)

See PaperTrail.VersionQueries.get_current_model/1.

get_model_id(model)

get_version(record)

See PaperTrail.VersionQueries.get_version/1.

get_version(model_or_record, id_or_options)

See PaperTrail.VersionQueries.get_version/2.

get_version(model, id, options)

See PaperTrail.VersionQueries.get_version/3.

get_versions(record)

See PaperTrail.VersionQueries.get_versions/1.

get_versions(model_or_record, id_or_options)

See PaperTrail.VersionQueries.get_versions/2.

get_versions(model, id, options)

See PaperTrail.VersionQueries.get_versions/3.

insert(changeset, options \\ [origin: nil, meta: nil, originator: nil, prefix: nil, model_key: :model, version_key: :version])

Inserts a record to the database with a related version insertion in one transaction

insert!(changeset, options \\ [origin: nil, meta: nil, originator: nil, prefix: nil, model_key: :model, version_key: :version])

Same as insert/2 but returns only the model struct or raises if the changeset is invalid.

update(changeset, options \\ [origin: nil, meta: nil, originator: nil, prefix: nil])

Updates a record from the database with a related version insertion in one transaction

update!(changeset, options \\ [origin: nil, meta: nil, originator: nil, prefix: nil])

Same as update/2 but returns only the model struct or raises if the changeset is invalid.