paper_trail v0.7.4 PaperTrail.Version

Functions

changeset(model, params \\ :empty)

count()

Returns the count of all version records in the database

first()

Returns the first version record in the database by :inserted_at

last()

Returns the last version record in the database by :inserted_at