paper_trail v0.8.7 PaperTrail.Version

Functions

changeset(model, params \\ :empty)

count()

Returns the count of all version records in the database

count(options)

first()

Returns the first version record in the database by :inserted_at

first(options)

last()

Returns the last version record in the database by :inserted_at

last(options)