paper_trail v0.8.7 PaperTrail.VersionQueries

Functions

get_current_model(version)

Gets the current model record/struct of a version

get_version(record)

get_version(record :: Ecto.Schema.t()) :: Ecto.Query.t()

Gets the last version of a record.

get_version(model, id)

get_version(model :: module(), id :: pos_integer()) :: Ecto.Query.t()

get_version(record :: Ecto.Schema.t(), options :: []) :: Ecto.Query.t()

Gets the last version of a record.

A list of options is optional, so you can set for example the :prefix of the query, wich allows you to change between different tenants.

Usage example:

iex(1)> PaperTrail.VersionQueries.get_version(record, [prefix: "tenant_id"])

get_version(model, id, options)

get_version(model :: module(), id :: pos_integer(), options :: []) ::
  Ecto.Query.t()

Gets the last version of a record given its module reference and its id.

A list of options is optional, so you can set for example the :prefix of the query, wich allows you to change between different tenants.

Usage example:

iex(1)> PaperTrail.VersionQueries.get_version(ModelName, id, [prefix: "tenant_id"])

get_versions(record)

get_versions(record :: Ecto.Schema.t()) :: Ecto.Query.t()

Gets all the versions of a record.

get_versions(model, id)

get_versions(model :: module(), id :: pos_integer()) :: Ecto.Query.t()

get_versions(record :: Ecto.Schema.t(), options :: []) :: Ecto.Query.t()

Gets all the versions of a record.

A list of options is optional, so you can set for example the :prefix of the query, wich allows you to change between different tenants.

Usage example:

iex(1)> PaperTrail.VersionQueries.get_versions(record, [prefix: "tenant_id"])

get_versions(model, id, options)

get_versions(model :: module(), id :: pos_integer(), options :: []) ::
  Ecto.Query.t()

Gets all the versions of a record given a module and its id.

A list of options is optional, so you can set for example the :prefix of the query, wich allows you to change between different tenants.

Usage example:

iex(1)> PaperTrail.VersionQueries.get_versions(ModelName, id, [prefix: "tenant_id"])